Making version is Long instead of Int and now we are supporting up to 6 digits in Major, minor & micro version instead of up to 2 digits.

+++ trunk/MetadataAPI/src/main/scala/com/ligadata/MetadataAPI/TestMetadataAPI.scala
@@ -180,7 +180,7 @@ object TestMetadataAPI{
       val typNameSpace = typKeyTokens(0)
       val typName = typKeyTokens(1)
       val typVersion = typKeyTokens(2)
-      val apiResult = MetadataAPIImpl.RemoveType(typNameSpace,typName,typVersion.toInt)
+      val apiResult = MetadataAPIImpl.RemoveType(typNameSpace,typName,typVersion.toLong)
 
       val (statusCode1,resultData1) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -292,7 +292,7 @@ object TestMetadataAPI{
       val fcnNameSpace = fcnKeyTokens(0)
       val fcnName = fcnKeyTokens(1)
       val fcnVersion = fcnKeyTokens(2)
-      val apiResult = MetadataAPIImpl.RemoveFunction(fcnNameSpace, fcnName, fcnVersion.toInt)
+      val apiResult = MetadataAPIImpl.RemoveFunction(fcnNameSpace, fcnName, fcnVersion.toLong)
 
       val (statusCode1, resultData1) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -436,7 +436,7 @@ object TestMetadataAPI{
       val msgName = msgKeyTokens(1)
       val msgVersion = msgKeyTokens(2)
 
-      val depModels = MetadataAPIImpl.GetDependentModels(msgNameSpace,msgName,msgVersion.toInt)
+      val depModels = MetadataAPIImpl.GetDependentModels(msgNameSpace,msgName,msgVersion.toLong)
       logger.trace("DependentModels => " + depModels)
 
       val apiResult = MetadataAPIImpl.GetMessageDef(msgNameSpace,msgName,"JSON",msgVersion)
@@ -481,7 +481,7 @@ object TestMetadataAPI{
       val msgName = msgKeyTokens(1)
       val msgVersion = msgKeyTokens(2)
 
-      val depModels = MetadataAPIImpl.GetDependentModels(msgNameSpace,msgName,msgVersion.toInt)
+      val depModels = MetadataAPIImpl.GetDependentModels(msgNameSpace,msgName,msgVersion.toLong)
       if( depModels.length > 0 ){
 	depModels.foreach(mod => {
 	  logger.trace("DependentModel => " + mod.FullNameWithVer)
@@ -658,7 +658,7 @@ object TestMetadataAPI{
       val msgNameSpace = msgKeyTokens(0)
       val msgName = msgKeyTokens(1)
       val msgVersion = msgKeyTokens(2)
-      val apiResult = MetadataAPIImpl.RemoveMessage(msgNameSpace,msgName,msgVersion.toInt)
+      val apiResult = MetadataAPIImpl.RemoveMessage(msgNameSpace,msgName,msgVersion.toLong)
 
       val (statusCode,resultData) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -699,7 +699,7 @@ object TestMetadataAPI{
       val contNameSpace = contKeyTokens(0)
       val contName = contKeyTokens(1)
       val contVersion = contKeyTokens(2)
-      val apiResult = MetadataAPIImpl.RemoveContainer(contNameSpace,contName,contVersion.toInt)
+      val apiResult = MetadataAPIImpl.RemoveContainer(contNameSpace,contName,contVersion.toLong)
 
       val (statusCode,resultData) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -740,7 +740,7 @@ object TestMetadataAPI{
       val msgNameSpace = msgKeyTokens(0)
       val msgName = msgKeyTokens(1)
       val msgVersion = msgKeyTokens(2)
-      val apiResult = MetadataAPIImpl.RemoveMessage(msgNameSpace,msgName,msgVersion.toInt)
+      val apiResult = MetadataAPIImpl.RemoveMessage(msgNameSpace,msgName,msgVersion.toLong)
 
       val (statusCode,resultData) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -780,7 +780,7 @@ object TestMetadataAPI{
       val modNameSpace = modKeyTokens(0)
       val modName = modKeyTokens(1)
       val modVersion = modKeyTokens(2)
-      val apiResult = MetadataAPIImpl.RemoveModel(modNameSpace,modName,modVersion.toInt)
+      val apiResult = MetadataAPIImpl.RemoveModel(modNameSpace,modName,modVersion.toLong)
 
       val (statusCode,resultData) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -821,7 +821,7 @@ object TestMetadataAPI{
       val modNameSpace = modKeyTokens(0)
       val modName = modKeyTokens(1)
       val modVersion = modKeyTokens(2)
-      val apiResult = MetadataAPIImpl.DeactivateModel(modNameSpace,modName,modVersion.toInt)
+      val apiResult = MetadataAPIImpl.DeactivateModel(modNameSpace,modName,modVersion.toLong)
 
       val (statusCode,resultData) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -862,7 +862,7 @@ object TestMetadataAPI{
       val modNameSpace = modKeyTokens(0)
       val modName = modKeyTokens(1)
       val modVersion = modKeyTokens(2)
-      val apiResult = MetadataAPIImpl.ActivateModel(modNameSpace,modName,modVersion.toInt)
+      val apiResult = MetadataAPIImpl.ActivateModel(modNameSpace,modName,modVersion.toLong)
 
       val (statusCode,resultData) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
@@ -902,7 +902,7 @@ object TestMetadataAPI{
       val modNameSpace = modKeyTokens(0)
       val modName = modKeyTokens(1)
       val modVersion = modKeyTokens(2)
-      val apiResult = MetadataAPIImpl.RemoveModel(modNameSpace,modName,modVersion.toInt)
+      val apiResult = MetadataAPIImpl.RemoveModel(modNameSpace,modName,modVersion.toLong)
 
       val (statusCode,resultData) = MetadataAPIImpl.getApiResult(apiResult)
       println("Result as Json String => \n" + apiResult)
