+++ trunk/MetadataAPIService/src/main/scala/com/ligadata/metadataapiservice/DeactivateObjectsService.scala
@@ -58,7 +58,7 @@ class DeactivateObjectsService(requestContext: RequestContext) extends Actor {
 
     arg.ObjectType match {
       case "model" => {
- 	      MetadataAPIImpl.DeactivateModel(nameSpace,arg.Name,version.toInt).toString
+ 	      MetadataAPIImpl.DeactivateModel(nameSpace,arg.Name,version.toLong).toString
       }
       case _ => {
 	      new ApiResult(-1, APIName, "Deactivate/Activate on " + arg.ObjectType + " is not supported yet").toString
